DAK 1440: Implement one at a time Hi2Lo capability
The Hi2Lo algorithm can now run iterations one at a time and exit
gracefully without running the high-fidelity model (which is meant to be
run offline)

+++ src/NonDBayesCalibration.hpp
@@ -230,7 +230,7 @@ protected:
   unsigned short importCandFormat;
   /// maximum number of high-fidelity model runs to be used for adaptive
   /// Bayesian experimental design
-  size_t maxHifiEvals;
+  int maxHifiEvals;
 
   // settings specific to model discrepancy
